+++ bot/__main__.py
@@ -1,0 +1,31 @@
+import os
+import logging
+from telegram import Update
+from telegram.ext import Application
+
+from handlers import register_handlers
+
+# Set up logging
+logging.basicConfig(
+    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
+)
+# Set higher logging level for httpx to avoid all GET and POST requests being logged
+logging.getLogger("httpx").setLevel(logging.WARNING)
+
+logger = logging.getLogger(__name__)
+
+
+def main() -> None:
+    """Allow running a bot."""
+    # Create the Application and pass it your bot's token.
+    application = Application.builder().token(os.getenv('BOT_TOKEN')).build()
+
+    # Register all handlers
+    register_handlers(application)
+
+    # Run the bot until the user presses Ctrl-C
+    application.run_polling(allowed_updates=Update.ALL_TYPES)
+
+
+if __name__ == "__main__":
+    main()
