+++ bot/database/__init__.py
@@ -1,0 +1,11 @@
+from bot.database.database import Database
+from bot.database.user import User
+from bot.database.product import Product
+from bot.database.order import Order
+
+__all__ = [
+    'Database',
+    'User',
+    'Product',
+    'Order'
+]
