+++ bot/database/product.py
