+++ bot/roles/__init__.py
@@ -1,0 +1,9 @@
+from bot.roles.admin import Admin
+from bot.roles.barman import Barman
+from bot.roles.customer import Customer
+
+__all__ = [
+    'Admin',
+    'Barman',
+    'Customer'
+ ]
