+++ bot/roles/barman.py
@@ -1,28 +1,8 @@
-# Бармен имеет очередь заказов, он их принимает и отмечает выполненными.
-# Также может заказать что-то сам, либо стать заказчиком.
-# Здеся либо класс, либо просто написать логику для менюшек, какие-то функции только для бармена и прочее, прочее.
 import logging
-import datetime
-from typing import Tuple, Any
-
-from modules.customer import Customer
-from modules.database import Database
-from modules.product import Product
-from modules.order import Order
-from modules.user import User
-
-
 from telegram import ForceReply, Update, InlineKeyboardMarkup, InlineKeyboardButton
-from telegram.constants import ParseMode
-from telegram.ext import Application, CommandHandler, ContextTypes, MessageHandler, filters, CallbackContext
 
-# Тексты для меню
-
-
-# Тексты для кнопок
-
-
-# Создание клавиатур
+from bot.roles.customer import Customer
+from bot.database import Database, Order
 
 
 class Barman(Customer):
