--- src/main.py
@@ -1,46 +1,0 @@
-# мэйн - точка входа, в мэйне проиcходит аунтетифакция пользователя, инициализация хэндлеров...
-# пока тут просто шаблон
-import os
-import logging
-
-from modules.customer import Customer
-from modules.barman import Barman
-from modules.admin import Admin
-
-from modules.handlers import echo, start, help_command, menu, button_callbacks
-
-from telegram import ForceReply, Update
-from telegram.ext import Application, CommandHandler, ContextTypes, MessageHandler, filters, CallbackQueryHandler
-
-# Enable logging
-logging.basicConfig(
-    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
-)
-# set higher logging level for httpx to avoid all GET and POST requests being logged
-logging.getLogger("httpx").setLevel(logging.WARNING)
-
-logger = logging.getLogger(__name__)
-
-
-def main() -> None:
-    """Start the bot."""
-    # Create the Application and pass it your bot's token.
-    application = Application.builder().token(os.getenv('BOT_TOKEN')).build()
-
-    # on different commands - answer in Telegram
-    application.add_handler(CommandHandler("start", start))
-    application.add_handler(CommandHandler("help", help_command))
-    application.add_handler(CommandHandler("menu", menu))
-
-    # Register handler for inline buttons
-    application.add_handler(CallbackQueryHandler(button_callbacks))
-
-    # on non command i.e. message - echo the message on Telegram
-    application.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, echo))
-
-    # Run the bot until the user presses Ctrl-C
-    application.run_polling(allowed_updates=Update.ALL_TYPES)
-
-
-if __name__ == "__main__":
-    main()
